+++ egov/egov-egfweb/src/main/java/org/egov/egf/web/actions/report/ExpenseJournalVoucherPrintAction.java
@@ -176,7 +176,7 @@ public class ExpenseJournalVoucherPrintAction extends BaseFormAction {
     private void generateVoucherReportList() {
         if (voucher != null) {
             for (final CGeneralLedger vd : voucher.getGeneralledger())
-                if (BigDecimal.ZERO.compareTo(BigDecimal.valueOf(vd.getCreditAmount())) == 0) {
+                if (BigDecimal.ZERO.compareTo((vd.getCreditAmount())) == 0) {
                     final VoucherReport voucherReport = new VoucherReport(persistenceService, Integer.valueOf(voucher.getId()
                             .toString()), vd, egovCommon);
                     if (billRegistermis != null)
@@ -184,7 +184,7 @@ public class ExpenseJournalVoucherPrintAction extends BaseFormAction {
                     voucherReportList.add(voucherReport);
                 }
             for (final CGeneralLedger vd : voucher.getGeneralledger())
-                if (BigDecimal.ZERO.compareTo(BigDecimal.valueOf(vd.getDebitAmount())) == 0) {
+                if (BigDecimal.ZERO.compareTo((vd.getDebitAmount())) == 0) {
                     final VoucherReport voucherReport = new VoucherReport(persistenceService, Integer.valueOf(voucher.getId()
                             .toString()), vd, egovCommon);
                     if (billRegistermis != null)
